+++ src/main/java/nl/d2n/dao/result/UserUpdateCount.java
@@ -2,17 +2,17 @@ package nl.d2n.dao.result;
 
 public class UserUpdateCount {
 
-    private Integer id;
+    private Long id;
 
     private Integer reads;
     private Integer writes;
 
-    public UserUpdateCount(Integer id, Integer reads, Integer writes) {
+    public UserUpdateCount(Long id, Integer reads, Integer writes) {
         this.id = id;
         this.reads = reads;
         this.writes = writes;
     }
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
     public Integer getReads() {
