+++ src/main/java/nl/d2n/model/InboxAlert.java
@@ -6,14 +6,14 @@ public class InboxAlert {
 
     @Expose private Integer messages;
     @Expose private Integer invitations;
-    private Integer userid;
+    private Long userid;
     private String tempkey;
     private Long valid;
     @Expose private String oourl;
 
     public Integer getMessages() { return this.messages; }
     public Integer getInvitations() { return this.invitations; }
-    public Integer getUserId() { return this.userid; }
+    public Long getUserId() { return this.userid; }
     public String getTempkey() { return this.tempkey; }
     public Long getValid() { return this.valid; }
     public String getOourl() { return this.oourl; }
